feat: add app-metadata gist
