+++ src-tauri/src/services/downloader/mod.rs
@@ -1,9 +1,7 @@
 pub mod model;
 pub mod server;
-pub mod dictionary;
-mod updater;
+mod gist;
 
 pub use self::model::*;
 pub use self::server::*;
-pub use self::dictionary::*;
-pub use self::updater::*;
+pub use self::gist::*;
